feat: add knex pg connection check
